tried to organize
