+++ app/src/main/java/com/example/minstrom/MainActivity.kt
@@ -5,39 +5,30 @@ import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
-import androidx.navigation.compose.composable
-import com.example.minstrom.screen1elements.TaskBox
-import com.example.minstrom.ui.theme.MinstromTheme
-import com.google.firebase.Firebase
+import com.example.minstrom.appViewModels.AppViewModel
+import com.example.minstrom.data.user.User
+import com.example.minstrom.data.user.UserViewModel
 import java.time.LocalTime
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.minutes
-import kotlin.time.DurationUnit
-import kotlin.time.toDuration
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
         setContent {
-            val appViewModel:AppViewModel = AppViewModel()
+            val appViewModel: AppViewModel = AppViewModel()
             val navController = rememberNavController()
             Navhost(navController = navController,appViewModel = appViewModel)
         }
     }
 }
+
+
+//kode herunder er gemt for sjov og bliver ikke brugt<3
 
 //Henter alle users og printer dem (i system.out)
 @Composable
