+++ app/src/main/java/com/example/minstrom/data/device/DeviceTransferClass.kt
@@ -1,4 +1,4 @@
-package com.example.minstrom
+package com.example.minstrom.data.device
 
 data class DeviceTransferClass (
     var id:String = "",     //id er var fordi den bliver overwritten efter obj bliver lavet som verify-check
