+++ app/src/main/java/com/example/minstrom/screen1elements/Screen1.kt
@@ -1,6 +1,5 @@
-package com.example.minstrom
+package com.example.minstrom.screen1elements
 
-import android.util.Log
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -13,14 +12,8 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateListOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -28,12 +21,9 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.core.graphics.toColorInt
-import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
-import com.example.minstrom.screen1elements.BlueButton
-import com.example.minstrom.screen1elements.MenuButton
-import com.example.minstrom.screen1elements.TaskBox
-import com.example.minstrom.screen1elements.TextOnPage
+import com.example.minstrom.appViewModels.AppViewModel
+import com.example.minstrom.R
 
 @Composable
 fun Screen1(
