+++ app/src/main/java/com/example/minstrom/screen1elements/TaskBox.kt
@@ -7,11 +7,8 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Text
@@ -20,22 +17,21 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
-import com.example.minstrom.AppViewModel
-import com.example.minstrom.Device
+import com.example.minstrom.appViewModels.AppViewModel
+import com.example.minstrom.data.device.Device
 import com.example.minstrom.R
 
 //box til opgaver screen1
 @Composable
 fun TaskBox(
-        device: Device,
-        appViewModel: AppViewModel,
-        navController: NavController
+    device: Device,
+    appViewModel: AppViewModel,
+    navController: NavController
 ) {
     Box(
         modifier = Modifier
