+++ app/src/main/java/com/example/minstrom/screen2elements/SettingButton.kt
@@ -1,0 +1,83 @@
+package com.example.minstrom.screen2elements
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.ModalBottomSheet
+import androidx.compose.material3.Text
+import androidx.compose.material3.rememberModalBottomSheetState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun SettingButton(
+    text: String,
+    img: Int,
+    bottomSheetIndhold: @Composable ()-> Unit //modtager en composable med indholdet til bottomsheet
+) {
+    var showBottomSheet by remember { mutableStateOf(false) }
+    val sheetState = rememberModalBottomSheetState(
+        skipPartiallyExpanded = false,
+    )
+
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(12.dp)
+            .clip(RoundedCornerShape(16.dp))
+            .background(Color.White)
+            .border(1.dp, Color.LightGray, RoundedCornerShape(16.dp))
+            .padding(horizontal = 20.dp, vertical = 12.dp)
+            .clickable(onClick = { showBottomSheet = true }) //OnClick
+    ) {
+        Row(
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+
+            //image
+            Image(painter = painterResource(img),
+                contentDescription = null,
+                Modifier.size(40.dp))
+
+            Text(
+                text = text,
+                color = Color.Black,
+                fontSize = 16.sp
+            )
+        }
+
+        Column {
+            if (showBottomSheet) {
+                ModalBottomSheet(
+                    modifier = Modifier.fillMaxHeight(),
+                    sheetState = sheetState,
+                    onDismissRequest = { showBottomSheet = false }
+                ) {
+                    bottomSheetIndhold() //det reele indhold i bottomsheet kaldes her
+                }
+            }
+        }
+    }
+}
